+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add `Account.ConnectivityHtml()`
 - add type for every event type
 - use `deltachat-rpc-server.exe` as executable name for the RPC process
+- remove Account.CreateChat()
 
 ## v0.13.0